KernelCpuSpeedReader: Account for missing sysfs nodes
It is possible that after CPU/cluster hotplugging
the cpufreq stats sysfs nodes don't get recreated.

In that case we would be encountering a lot of failed
read attempts which are obsolete.

As solution, this adds a detection mechanism
which checks the cpu online status and implements
a failure counter.
The failure counter gets incremented every time
a read attempt to the stats sysfs node fails while
the cpu was online. This case is considered as proof for
the sysfs bug.
Once the failure counter reaches MAX_READ_TRIES (3 by default),
no more read attempts are made.

Test: Make otapackage
Test: Manual test which verified that no more read
Test: attempts are made after the limit was reached.
Test: Verified that the read from the cpu online node
Test: was successful.

Change-Id: Ia7c413c43d52757a310c9f6105a466dba381e9ed
Signed-off-by: Alex Naidis <alex.naidis@linux.com>